requirements.txt
Flask==3.1.3
Werkzeug==3.1.7
python-dotenv==1.1.0
requests==2.32.3
urllib3>=1.26.0,<2
APScheduler==3.11.2
gunicorn==23.0.0
flask-limiter==3.9.0
flask-wtf==1.2.2
reportlab==4.2.5
